add transformation for robot vision
